+++ example/lib/main.dart
@@ -29,7 +29,7 @@ class _PaymentViewState extends State<PaymentView> {
   @override
   void initState() {
     super.initState();
-    //! Initialize UniPay can be now also be done by calling this method
+    //! Initialize UniPay can be now also be done by calling this method, if you require to use `Apple Pay` only.
     // UniPayServices.initUniPay(uniPayData: uniPayData, context: context);
   }
 
@@ -39,10 +39,13 @@ class _PaymentViewState extends State<PaymentView> {
       context: context,
       uniPayData: uniPayData,
     );
+
+    // ----- Standalone Apple Pay view ----- //
     // return _applePayOnlyView();
   }
 
   // To show the Apple Pay only view
+  // ignore: unused_element
   Widget _applePayOnlyView() {
     return Scaffold(
       appBar: AppBar(title: const Text("UniPay Example")),
@@ -56,7 +59,7 @@ class _PaymentViewState extends State<PaymentView> {
   }
 }
 
-/// UniPayData to be used for payment request
+/// UniPayData to be used for payment request, make sure to fill all the required fields properly.
 UniPayData uniPayData = UniPayData(
   appName: "UniPay",
   locale: UniPayLocale.en,
@@ -65,7 +68,7 @@ UniPayData uniPayData = UniPayData(
     email: "example@mail.com",
     phoneNumber: "+966555666777",
     address: UniPayAddress(
-      addressName: "Olaya street, Al Ghadir",
+      addressName: "KAFD Area, Al Ghadir, Riyadh, Saudi Arabia",
       city: "Riyadh",
     ),
   ),
@@ -80,16 +83,17 @@ UniPayData uniPayData = UniPayData(
     ],
     moyasarCredential: MoyasarCredential(
       publishableKey: "pk_test",
-      secretKey: "sk_test",
+      secretKey: "sk_live",
     ),
     tamaraCredential: TamaraCredential(
-      token: "Bearer test",
+      token: "Bearer test_12345",
       merchantUrl: MerchantUrl(notification: "https://my-app.com/webhook"),
     ),
     tabbyCredential: TabbyCredential(
       psKey: "pk_test",
       secretKey: "sk_test",
       merchantCode: "your_merchant_code",
+      merchantUrl: MerchantUrl(notification: "https://my-app.com/webhook"),
     ),
   ),
   orderInfo: UniPayOrder(
@@ -112,7 +116,7 @@ UniPayData uniPayData = UniPayData(
     debugPrint("Payment Failed ----> ${res.toMap()}");
   },
   metaData: {
-    "customerId": "ABC_12345",
-    "customerName": "Mohammad Saif",
+    "customer_uid": "ABC_12345",
+    "customer_name": "Mohammad Saif",
   },
 );